trying different approach on TestRPCStorageHandler_Disconnect
